Update patch set 4
Patch Set 4:

Since the code is quite similar to the already merged ClassMembersInitializationChecker I looked at its code. It seems you hit a bug indeed. This code doesn't work:

class Waldo {
	int location;

	Waldo() {
		Waldo d;
		d.location = 1;
	}
};

There isn't warning about lack of field initialization because the checker thinks that "location" is the field of the class but in this
case we are initializing another instance of the same class.

Patch-set: 4
Label: Verified=0