Update patch set 3
Patch Set 3: Code-Review-1

Umair, I can see that this change fixes the local problem, but I am a lot less confident that it doesn't introduce other issues.

The change you introduce is essentially the same(ish) as removing this bit of code in org.eclipse.cdt.core.resources.ACBuilder.addMarker(ProblemMarkerInfo)

			// Set source attribute only if the marker is being set to a file from different project
			if (project != null && !project.equals(markerResource.getProject())) {
				marker.setAttribute(IMarker.SOURCE_ID, project.getName());
			}


which has the effect of org.eclipse.cdt.internal.core.BuildRunnerHelper.removeOldMarkers(IResource, IProgressMonitor) not removing the markers that were created. That seems to me important code and this change may cause a regression of Bug 105978. 

However, the biggest issue I see here (and possibly the underlying issue) is that library project is built twice.


I am not ready to say 100% no to this patch, please review my findings and comment.

Patch-set: 3