Update patch set 3
Patch Set 3:

I am in favour of approving this despite John's comment. However it would be best to fix the known remaining issues. 

The reason I am in favour is that there will be an unending list of places that this is kind of thing is a problem, so I think that if the change does not break existing users and use cases then we should open up the additional paths. Advantest have been doing a fair number of contributions back so making giving a path seems preferable than not.

The reality is that GnuMakefileGenerator and its surrounding code is mostly unmaintained, CDT's intention is to move to CDT core build. 

John, WDYT?

Patch-set: 3