Update patch set 3
Patch Set 3:

Thanks Jonah for the review! See the responses below.

 > The change you introduce is essentially the same(ish) as removing
 > this bit of code in org.eclipse.cdt.core.resources.ACBuilder.addMarker(ProblemMarkerInfo)
 > 
 > // Set source attribute only if the marker is being set to a file
 > from different project
 > if (project != null && !project.equals(markerResource.getProject()))
 > {
 > marker.setAttribute(IMarker.SOURCE_ID, project.getName());
 > }
 > 
 > 
 > which has the effect of org.eclipse.cdt.internal.core.BuildRunnerHelper.removeOldMarkers(IResource,
 > IProgressMonitor) not removing the markers that were created. That
 > seems to me important code and this change may cause a regression
 > of Bug 105978.

There are two scenarios.
1. A project is using a file from some other project and it is built with current project instead of its own project, like inclusion of header.

2. A project is referencing some other project (say B). Triggering build on current project triggers own build of project B and all of project B files are built with the project B build.

Bug 105978 says: "If you have a project A, which has a *included file* (i.e. header) located in a simple project B(not C/C project) in the same workspace, any error markers on files in Project B will not be removed even you correct the error and rebuild the project A."

So it is definitely the case 1. And in this case, condition "!project.equals(markerResource.getProject()" returns true, so the current change is not like removing this piece of code from ACBuilder.addMarker(ProblemMarkerInfo). I have verified that case 1 works fine. In case 1, IMarker.SOURCE_ID is the current project for markers and BuildRunnerHelper.removeOldMarkers(..) removes all such markers that are not in current project but the markers' source is current project.

The current patch is for the case 2. Each *project build* keeps its own markers; either build is invoked directly or indirectly (referenced by some other project).

 > However, the biggest issue I see here (and possibly the underlying
 > issue) is that library project is built twice.

This is another problem; I'll definitely open a bug for this sometime :) The underlying problem is that in CDT projects, we have two type of references; eclipse project type reference and the CDT type references. When we add a reference to CDT project from "Paths and Symbols > References" tab, it adds both of the references. In eclipse type project references, AFAIK there is no concept of configurations.

When we trigger a build on CDT project, eclipse build system finds the eclipse type project reference and triggers the build on it and that referenced project is built with active configuration. The CommonBuilder again gets the referenced projects and referenced build configs and triggers builds on them. I set the project reference to Release config of lib project. The two builds you are seeing are; one that is Active config (Debug) built by eclipse core build system and the other one and the correct one Release config built by CommonBuilder.

We should think of removing the eclipse type project reference to avoid unnecessary build invocations in case of project references chain. The other problem with this eclipse type project reference is that if I clean a project then it cleans all the referenced project which is not desired.

If you say, I can open a bug right away (if there is no existing bug for it) and keep the discussion going there. WDYS?

 > 
 > 
 > I am not ready to say 100% no to this patch, please review my
 > findings and comment.

Just let me know any other concerns and I'll investigate :)

Patch-set: 3
Label: Verified=0