Update patch set 4
Patch Set 4:

(1 comment)

Thanks! I'm fine with this approach, i.e. I don't think we need to go making changes to fAssociatedValue unless we have other scenarios to motivate it.

The commit message makes it sound like we are only expanding the set of variables treated as constant expressions, but in fact we are also restricting it (by no longer including variables which are non-const but have an initial value which is a constant expression); please clarify that.

Patch-set: 4
Label: Verified=0