Update patch set 3
Patch Set 3:

> > 1. Add ICPPEvaluation ICPPFunctionType.getNoexceptSpecifier().
 > 
 > Should I add ICPPASTExpression instead of ICPPEvaluation as the
 > latter is not API?

Making it ICPPASTExpression would defeat the purpose, as we need this to return a result even for index bindings (where there is no AST available).

If we wanted to make it public API, we could make it return an IValue. (That can then be turned into an ICPPEvaluation as follows: if it's a DependentValue, return the wrapped evaluation. Otherwise, construct an EvalFixed wrapping the value.)

However, I think it would be simpler to just make it ICPPEvaluation, and mark it as @noreference for API purposes (similar to ICPPASTInitializerClause.getEvaluation()). After all, the callers we want to write will be internal. We can always turn it into a public API later if there is demand for that.

Patch-set: 3
Label: Verified=0