Update patch set 8
Patch Set 8:

I tried again to get rid of the inCalledContext.

Did you have a look at the example that I posted as an example why I needed it (in the comment on patch set 4)?

The problem there is that the fArguments[0] of the EvalFunctionCall is an EvalUnary which has an EvalBinary and somewhere there the CPPVariable for the member function ptr is hidden. Therefore, I have to somehow recurse into the Evaluations. Any ideas?

Patch-set: 8