Update patch set 1
Patch Set 1:

If I understood the code this value is true if the system header option is not selected for the entry from the provider. When the built-in cdt provider is used the user can't select anything, everything is added to the include path settings is considered as system header. I can see if it's possible to perform this check at provider level in order to get the correct result from isForQuoteIncludeOnly.

Patch-set: 1