Update patch set 1
Patch Set 1:

I read again your comment, but maybe I'm missing something. The method isForQuoteIncludesOnly just returns what we used in the constructor and it's exactly what we are doing, during construction of an IncludeSearchPathElement we set if the include is for quote only or not. Please clarify because as I said maybe I'm missing something.

Patch-set: 1
Label: Verified=0