Update patch set 1
Patch Set 1:

Hi Marco, take a look at this version which uses a TextEditorAction and can determine if enabled or not based on what is selected.  Your version has a flaw in that enablement only sees the method granularity (e.g. that selection is in main()) and not what was actually selected.  This is because the Outline view ends up calling the selection changed listener with the function/method instead of the low-level element.  This version could be modified to not disable and simply store whatever text was selected if item does not have a qualified name.

Patch-set: 1