Update patch set 1
Patch Set 1:

I'm sorry but I'm not able to reproduce what you are saying with the latest version of my patch. If I select a global variable for example out of any method from the text editor, the menu is enabled. If there's no selection the menu is disabled. If I select something from the outline view is enabled again and so on.

Patch-set: 1