Update patch set 1
Patch Set 1:

With your version of the patch and the following code:

#include <iostream>
using namespace std;
int main(int argc, char **argv) {
	cout << "hello world" << endl;
	return 0;
}

Highlight return and go to source menu.  Copy qualified name will be enabled.  It will then issue the error message when clicked or CTRL-Q is sent.  When I debugged, it is passing the Function main to the validate routine which succeeds.  This is due to the Outline view changing the ITextSelection into it's own granularity and then calling your validate which can't tell what the original selection was anymore.  With the text version, you get the original text selection to validate.

Patch-set: 1