Update patch set 1
Patch Set 1:

No problem.  I didn't know this section of the code until I tried it.  I looked at the JDT's version.  They leave it enabled and give the error message.  They also underline a letter in the name but that mnemonic doesn't work.  I figured the Copy logic seemed reasonable.  It is disabled when nothing selected and it does nothing when you use the CTRL+C and nothing is selected.  Thanks for the contributions.

Patch-set: 1