Update patch set 1
Patch Set 1:

Jeff, after your comment I can see a significant performance drop during text selection, at least on one of my test pc. I reverted the commit and the speed of text selection is ok again. I think all the work done in "isValidateSelection" needs to be removed with the exception of the first line. The method is correct but it's too heavy to create and parse the AST every time there's a selection. Can you perform a small commit with the fix? Thanks.

Patch-set: 1