Update patch set 1
Patch Set 1:

I didn't manage to write a test for my example. I tried to follow the patterns in FaultToleranceTests, but couldn't make the following code to fail (which triggers the StackOverflowException if I have it in the editor.

struct A {};

struct B{
	B(A) {
		C 
	}
}; 

Additionally, I am wondering if this is a good fix (for sure it doesn't harm). Maybe I should investigate why we actually end up in the aggregate init for this code snippet, because there is no aggregate init involved... I have absolutely no background on how to parse incomplete code.

Patch-set: 1
Label: Verified=0