Update patch set 7
Patch Set 7:

> It seems there's a typo in your example. You wrote Copyrsight
 > instead of Copyright. I used the regex '//=.*//\\s*Copyright.*' and
 > it works for me.

Sorry, I must have copy-pasted after a few experiments. But it still doesn't work for me.
I'm not sure how it would work as I am stepping through the code and comment.getRawSignature() only fetches the first line of the comment so it is not possible to match the 'Copyright' string.

Patch-set: 7