Update patch set 7
Patch Set 7:

ok, I think there was a misunderstanding, I tried the regex with a stub, at the moment my workspace is a bit a mess :) Ok, so basically the problem is not the regex but the definition of "first comment". Since they are all separated comments the check fails. I think we should add an option to cover this case.

Patch-set: 7
Label: Verified=0