Fixes for Test Leaking Editor test
The testLeakingInstanceAfterClose test relied on some weird
behaviour of an earlier test (testScalabilityDialogNotDismissedInadvertently_417909)
which had the effect of opening the editor window without the
welcome screen. This change makes that explicit and adds a little
bit of logging to identify why the testLeakingInstanceAfterClose
test may sometimes fail.

I have also increased the maximum time allowed before the test timesout.

Change-Id: I3433ccf1fc02cff76eeb278d05fee082157ca49c