Bug 547831 Allow LC Providers to use the Null Target
The Null Target is half way to becoming a first class citizen. However
we don't provide it in the list of targets available for a launch
descriptor. This adds support for that by adding a boolean flag to
the launch configuration provider launch bar contribution extension.

To help with that, changed the handling of get launch config for the
default launch descriptors to grab the config directly from the
descriptor without going through the providers. It would be very weird
for the providers to return a different config, but that door is now
shut.

Also add some documentation to ILaunchBarManager. Was going to add
the support there but went for the extension instead. Might as well
check in the docs I wrote doing that.

Change-Id: Ia03002a661a3971df68f74b2c338fe538b8b376a