Bug 547831 Add ILaunchTargets for GDB Remote
Adds ILaunchTargets for GDB Remote for TCP and Serial Port.
Adds a launch config provider that maps default launch configurations
for the gdb remote launch config types.

Alters the launch attributes by merging in the target attributes
over the launch configuration attributes. This is a no-op of the
target is the Null Target (which has no attributes).

Some string externalization that was missed from previous commits.

Carrying on tradition of ramdonly using Gdb or GDB in our class names :)

Change-Id: Ie8483110f594db593e704adda420ce6b14812dea