Update patch set 1
Patch Set 1:

Hi Jonah, I found this problem doing some testing on the latest CDT master branch.  I had seen the problem appear before randomly but I figured it out by adding a traceback to see it was during finalization of the DockerConnection.  I made a change in Linux Tools and added the CDT code using reflection as I realize you can't get a build of Linux Tools to use for RC2.  Hopefully you are ok with it.

Patch-set: 1