Update patch set 1
Patch Set 1:

After testing with the new Linux Tools fix, this change is no longer necessary.  The synchronized keyword is not accomplishing anything because it is synchronizing on the Job which changes from event to event.  I will play with adding a common object that the change events can synchronize with, but it will have to be for 9.9 as adding synchronization this close to RC2 is not a good idea.

Patch-set: 1
Reviewer: Gerrit User 582 <582@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0