Update patch set 1
Patch Set 1: Code-Review-2

Hi Marco,

Is this a critical item for 9.8.0? We have already made our final release candidate so only show-stoppers should be considered at this point.

I have realized that I did not mention this and will email to cdt-dev momentarily, but cdt_9_8 branch is closed to all but show-stoppers until after the release next week.

My -2 is to make sure it is not submitted too early, rather than a comment on the change itself which looks fine.

Jonah

Patch-set: 1