Bug 548141 - Fix Codan project options
The description of project options was misleading and it was't
really clear for the user what was going on if he enabled the
build options in the Codan preferences without enabling the
build option in the project. Keep now the project build option
enabled so the user can tune the option for each checker and
optionally he can turn off build check for all checkers using
the project option.

Change-Id: I8001a638468a21e5b8ae8b19070d757f28dd79df
Signed-off-by: Marco Stornelli <marco.stornelli@gmail.com>