Update patch set 6
Patch Set 6:

I'm not 100% sure about this. I guess there is a bit of matter of taste here but I would use only quotes for same folder headers or subfolders (rarely) and angle brackets for anything that is on the include path.
The behavior of quotes vs angle brackets is also not the same for every compiler. For example, MSVC checks for "quote includes" relative to the current file then if not found checks relatively to each previously included file (in reverse order of being opened). So in that context I would use quotes more sparingly.
It's also not uncommon to have many "external" dependencies within the project folder and those would not quality as "defined by us". In other words, I don't think checking if the header is contained with the project path is a good enough indicator of "defined by us".

One thing that's interesting is that if you add the include paths through the "Preprocessor Include paths, Macros etc" page, you can select wether or not an include path is considered system (angle brackets) or not ("quotes"). Perhaps the build "Setting" page could have a checkbox like that.

Patch-set: 6
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=0