Update patch set 17
Patch Set 17:

(20 comments)

Apologies for the long delay here. I've looked at most of the core changes so far. Haven't looked at rewrite or tests yet.

A high-level comment: it would be nice to cache the type and value of the initializer expression somewhere, rather than recomputing it once for every binding.

A suggestion for how this could be done:

* Make ICPPASTStructuredBindingDeclaration an IASTImplicitNameOwner, and give it an implicit name which resolves to a CPPVariable representing the complete object being destructured. This variable would cache its type and initial value like all variables.

* The individual bindings can then get this CPPVariable from the declaration, and do the additional processing specific to the binding.

Let me know if that suggestion makes sense.

Patch-set: 17