Update patch set 17
Patch Set 17:

(30 comments)

> Patch Set 17:
> 
> (11 comments)
> 
> Ok, done with the review now.
> 
> Thanks a lot for working on this and once again apologies for the delay.

Thanks a lot for your review. I fixed as much as I could. There are a few questions left. The main issue is that ExecIncomplete Objects cannot be marshalled, which leads to issues when having evaluations result in FixedEval.INCOMPLETE. But I think this is an issue beyond the implementation of structured bindings.

Adding an implicit name to the structured binding for the initializer sounds good to me. I'll approach this issue next week. However, I'll submit my progress to gerrit, since I already changed substantial parts of the code.

Patch-set: 17