Bug 549014 - Manage CMake toolchain files by toolchain
Use the toolchain type and id as the key to the Map holding the
toolchain files instead of the Path to the file. Also simplifies the
build config since we can always figure out the toolchain file from
the toolchain. Deprecate getting toolchain by path since that shouldn't
be needed any more.

Also took the opportunity to add javadoc to the interfaces.

Change-Id: I11ae2ad8177a3f60399742c8c19576f85ea4b8c0