Update patch set 5
Patch Set 5:

(4 comments)

Apologies for the long delay in getting this reviewed.

An alternative parsing strategy which would avoid additional backtracking would be: if we parsed a t_unspecified decl-specifier and a trailing return type (i.e. in the place where the patch adds a backtrack), check if the requirements for a deduction guide are met (i.e. the only decl specifier present is 'explicit', there are no nested declarators, the trailing-return-type has the form of a template-id, etc.), and if so then synthesize a CPPASTDeductionGuide from the relevant already-parsed nodes (otherwise backtrack). WDYT?

Btw. thanks for the notes in http://master.arknet.ch/~sophia/deduction_guides/, they were very helpful!

Patch-set: 5
Label: Verified=0