Update patch set 6
Patch Set 6:

Unless I rewrite totally the code for this checker I can't see a legal way to insert this code into the Eclipse baseline. I'm not a lawyer, the author wrote other accepted code for Eclipse and I think the ELA was active when the code was posted, but I can't say that it's now ok or not.

Patch-set: 6