Update patch set 7
Patch Set 7:

(4 comments)

Thanks, this is looking better!

I have one concern about the usability of this: it's really easy for this to trigger bogus diagnostics if there are ProblemBindings in the method body.

For example:

  struct Waldo {
    void foo(int);
    void bar() {
      foo(2, 3);  // wrong number of arguments
    }
  };
  
In addition to the overload resolution error on foo(), we'll get a "bar() should be static", because "foo" resolved to a ProblemBinding, which is not one of the class's methods.

To mitigate this, what do you think about suppressing these diagnostics if any field or method reference inside the method body resolves to a ProblemBinding? We can implement this by:

* having a flag "hasProblems" in ContextInfo, which is cleared in setMethodInfo()
* setting the flag in isClassField() and isClassMethod() if the name resolves to a ProblemBinding
* suppressing the diagnostics in leave(IASTDeclaration) if the flag is set

Patch-set: 7