Update patch set 20
Patch Set 20:

(4 comments)

I've drafted the caching implementation for the initializer. I considered the suggested approach using a CPPVariable. This seeming unfitting to me, as a CPPVariable is bound to an IASTName. I couldn't figure out a reasonable association of the imaginary object represented by the initializer of a structured binding. Therefore, I tried another approach. I extended the CPPASTStructuredBinding by a name evaluator strategy. It contains implementations for arrays, tuple-like decomposition, member decomposition and a constant of undecidable cases. The strategy is evaluated lazily and stored in the CPPASTStructuredBinding. In case of the resolution of a CPPVariable associated with a structured bindings name, the corresponding strategy is used.
In my opinion the caching needs to happen in the CPPASTStructuredBinding anyway and using CPPVariable seems coerced to me. What I don't really like in the current implementation that it needs access to the concrete CPPASTStructuredBinding class as I'd rather avoid putting the name resolution strategy interface to the public API through ICPPASTStructuredBinding.

Any thoughts on this?

Patch-set: 20
Label: Verified=0