Update patch set 8
Patch Set 8:

I also found some cases where the checker produces diagnostics which are not technically false positives, but are still undesirable.

For the static checker, it triggered a lot on "function object" types:

    struct A {
        // no fields

        template <typename T>
        T operator()(T t) {
            // ...
        }
    };

Here, this structure is intended to be used as a function object. While the call operator could be made static and the structure itself would still compile, but call sites that try to use is as a function object would not.

I think we should just except operator() from the "static" analysis.

Patch-set: 8