Update patch set 8
Patch Set 8:

For the const checker, I ran into cases like this:

    struct Foo {};
    struct B {
        Foo& foo;
        
        const Foo& get_foo() const { return foo; }

        // "get_foo" should be constant
        Foo& get_foo() { return foo; }
    };

Technically, the diagnostic is accurate: in isolation, the second overload would compile if made 'const'. However, the class as a whole would not compile (because now the two overloads have the same signature). Moreover, conceptually, exposing a non-const reference member is conceptually a modifying operation, even if the language doesn't enforce that.

Similarly:

    struct Impl {
        void foo();  // not constant
    };
    struct C {
        Impl* impl;
        
        // "foo should be constant"
        void foo() {
            impl->foo();
        }
    };

Again, this will compile if C:foo() is made const. But conceptually, calling a non-const function on a pointer member is conceptually a modifying operation.

Patch-set: 8
Label: Verified=0