Update patch set 8
Patch Set 8:

Here a set of possible strategies to cover these cases:
1) We can check for default and deleted methods, quite easy;
2) Don't check for "static" if the method is an overload operator, like operator() and so on;
3) Don't check for const if the method is returning a no-const reference.

In order the latest case from Nathan: calling a no-const method using a pointer is a modifying operation basically but the current code is deeply based on the method CPPVariableReadWriteFlags.getReadWriteFlags and in this case there's no write operation. Are we saying there's anything wrong with this method?

Patch-set: 8