Update patch set 12
Patch Set 12:

Maybe too much :) All you posted is direct "child" of what VariableReadWriteFlags "see" since the checker is deeply based on that to evaluate if a method can be const or not. We can do something to improve it but as in the last example you posted it would be a bit challenging to perform these "indirect" checks because we need to track the flow of code in someway. I will take a look but I won't be able to post codes in the next two weeks.

Patch-set: 12
Label: Verified=0