Update patch set 13
Patch Set 13:

As for tracking the flow of code... here are some thoughts on that.

First, observe that the following does not compile:

  struct A {
    int arr[10];

    void foo() const {
      int& x = arr[0];
      // we never assign to x
    }
  };

That is, the const-correctness violation occurs as soon as you take a non-const reference to a member. It doesn't require assigning to it. So, in cases like this, there is no need to track "flow".

However, we _can_ devise a testcase where the const-correctness violation is "far away" from the member access:

  struct A {
    int arr[10];

    void foo() const {
      auto&& x = arr[0];
      int& y = x;
    }
  };

Here, the deduced type of "x" is "reference to const int", and it's on the next line that we get an error about trying to convert it to "reference to int".

This last example suggests that trying to check const-correctness by only looking at member accesses will not work in general. We need to check other things, like type-checking the initialization of variables, throughout the function body.

It also suggests that for the "Method should be constant" check, to catch all cases we'd need an implementation strategy similar to this:

* Assume we have a procedure to fully type-check a function body (that is, run checks that would diagnose existing "Class member cannot be written" errors as well as cases like the above where an initialization fails).
* First, run the procedure on the method as written (non-const). If it gives an error, bail. (We are approximating this today with the "hasProblems".)
* Then, temporarily change the method to be const, and run the procedure to type-check it again. If that doesn't produce any errors, we know it's safe to issue the "Method should be constant" warning.

Patch-set: 13