Update patch set 13
Patch Set 13:

By the way, in case it's not clear, it's not just initialization. Adding or removing a "const" can potentially cause just about any category of error to appear. It doesn't even have to be an error directly in the function body - it could be in a template instantiation triggered by the function body.

For example:

    void foo(int&);
    
    template <typename T>
    void bar(T& t) {
	foo(t);
    }
    
    struct A {
	int field[10];
	
	void method() {
	    auto&& x = field[0];
	    bar(x);
	}
    };

Here, making method() const (as the checker suggests) introduces a compiler error, but it's inside the instantiation of bar().

Patch-set: 13
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0