Update patch set 2
Patch Set 2:

(3 comments)

Hi Adrien, I like that the change is shorter - but I didn't intend for you to go rewrite everything yet :-) 

AFAICT the 64 version can do both 32 and 64 bit files with the optional headers? If so, does that mean we don't need the original versions anymore?

For example, you mentioned that if it wasn't for API changes you could just modify the original to add support if I understood you correctly?

Patch-set: 2