Update patch set 2
Patch Set 2:

While I am currently a maintainer of CDT, I am less familiar with this area of code and AFAIK none of the original authors of this section are still active. 

Looking around the code base I can see that a similar issue was encountered with MachO, and in that case the MachO64 was a copy of MachO and that MachO was supposed to have been deleted, just no one remembered to.

The reason I bring this up, is that while I don't like copied code, renamed classes (even with a long time between the copy and the delete) is ok. The first patch is, as a result, a cleaner bit of code - so perhaps we should use that copy and deprecate for deletion the older parser. 

Thoughts?

Patch-set: 2
Label: Verified=0