Update patch set 3
Patch Set 3:

> While I am currently a maintainer of CDT, I am less familiar with
 > this area of code and AFAIK none of the original authors of this
 > section are still active.
 > 
 > Looking around the code base I can see that a similar issue was
 > encountered with MachO, and in that case the MachO64 was a copy of
 > MachO and that MachO was supposed to have been deleted, just no one
 > remembered to.
 
I remember! just very low priority :) The thing is that we can't just delete it, we need to do an upgrade path for people with a old project still referencing the old one. We'll need to do that for PE64 too. I'm not familiar with project conversions/upgrades but I might have to look into this soon for bug 360919 which will involve changing the MSVC toolchain to have it's own language instead of GNU C/C++ and therefore a bunch of stuff need to be migrated. Let me know if you are aware of a mechanism to do this.

Patch-set: 3