Update patch set 8
Patch Set 8:

I'm not agree, float comparison is a source of several errors, sometimes hard to find regardless the specific app domain. However we are not forcing any policy here because the checker will be disabled by default. It's a user choice to enable this kind of error check.

Patch-set: 8