Update patch set 7
Patch Set 7:

It's a general purpose checker to find not allowed methods/function. There are several rules used by companies to avoid for example sprintf &co. in c++ or find_first_of of string class. To be noted that the search is performed taking into account the fully qualified name.

Patch-set: 7