[releng] Bump version number of docs plug-ins
These docs plug-ins were having baselinereplace-p2-metadata. This is due
to still having problems with the compare&replace logic. The root pom.xml
lits <pattern>index/*</pattern> to exclude from compare, which means if
the only thing that changes is the generate files, and not the source
files, then the jar is a candidate for replace. In commit 3bab315fd23,
the version of help-docs-eclipserun-repo was changed, which means that
the generated output changes too.

Change-Id: I7f73159776d209aadaa70587988652e125b60851