Update patch set 3
Patch Set 3: Code-Review+1

Thanks. I would suggest going with this fix for now.

We can keep working on https://git.eclipse.org/r/#/c/149212/ as well if you'd like, but that's a riskier change, and in case we have to back it out it's good to have this fix to fall back on. (I also unfortunately don't have much time to deal with fallout from regressions these days.)

Patch-set: 3