Update patch set 8
Patch Set 8: Code-Review-1

(2 comments)

Is it worth going to the trouble of checking for things like "a >= 3.14 && a <= 3.14"? Have you ever seen someone write code like that in a real codebase?

Or are you worried that people will write this to work around a warning for "a == 3.14"?

Patch-set: 8
Label: Code-Review=-1