Update patch set 2
Patch Set 2:

The behaviour of the wizard is no different with this change, just the modal dialog is displayed - reality is I wouldn't trust the state of the wizard after one of these exceptions, but at least a user can be given a chance to correct the flow.

Note, other wizards (such as the tools.templates based ones and the PDE feature project wizard that I based this change on) actually close and then display the error pop-up.

Patch-set: 2