Update patch set 2
Patch Set 2:

(1 comment)

> > Uploaded patch set 2.
 > 
 > Ooops, there are more things to put in order between .target and
 > .setup, need to upload another Patch Set

OK. Looking forward to it. The cdt-baseline.target seems much shorter, AFAICT that is down to the source features being left out which I think is fine?

The cdt.target file (next to cdt-baseline.target) also need synchronizing.

As CDT 9.10.0 has just been released, one of my jobs is to update these files to use 9.10 as the baseline and 4.14 as the target platform, but rather than causing conflicts I will let your work get in first. Feel free to include those updates as part of this.

Patch-set: 2