Update patch set 9
Patch Set 9:

> I am +2 on this - except for the .target file contents being
 > reordered. Is it a tool or automatation that did this? If so, it
 > will happen again and again so I will live with it. If not and it
 > is possible to preserve order that would be great.

Typically I'm editing .target in its "source" form, because PDE tooling is far from the perfect state and I do not have resources for today to rewrite it.

I beleieve that order of dependencies does matter. This is why I normalized it a bit to support "upstream to downstream" order.

And you are right, it should not be a part of this change, so I reverted "order change" for cdt.target. For cdt-baseline.target I would like to keep it, because now it is much more readable.

Patch-set: 9
Reviewer: Gerrit User 23745 <23745@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0